requirements.txt
catboost==1.0.6
click==8.1.3
cycler==0.11.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.2.2
fonttools==4.36.0
graphviz==0.20.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
plotly==5.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
tenacity==8.0.1
threadpoolctl==3.1.0
Werkzeug==2.2.2
